move apollo specific code to apollo-compiler
