+++ libraries/apollo-compiler/src/test/kotlin/com/apollographql/apollo3/compiler/TypenameTest.kt
@@ -3,11 +3,9 @@ package com.apollographql.apollo3.compiler
 import com.apollographql.apollo3.ast.GQLDocument
 import com.apollographql.apollo3.ast.GQLFragmentDefinition
 import com.apollographql.apollo3.ast.GQLOperationDefinition
-import com.apollographql.apollo3.ast.builtinDefinitions
+import com.apollographql.apollo3.ast.introspection.toSchema
 import com.apollographql.apollo3.ast.toExecutableDefinitions
 import com.apollographql.apollo3.ast.toUtf8
-import com.apollographql.apollo3.ast.transformation.addRequiredFields
-import com.apollographql.apollo3.ast.introspection.toSchema
 import com.google.common.truth.Truth.assertThat
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
@@ -17,7 +15,6 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import java.io.File
 
-@Suppress("UNUSED_PARAMETER")
 @RunWith(TestParameterInjector::class)
 class TypenameTest(
     @TestParameter(valuesProvider = GraphQLFileValuesProvider::class) private val graphQLFile: File,
