+++ libraries/apollo-compiler/src/test/kotlin/com/apollographql/apollo3/compiler/keyfields/operations.graphql
